matchparen plugin use the <restore> attribute to :windo
This should prevent nasty side effects in other plugins that make use of
the matchparen plugin (e.g. the LargeFile) plugin, which do not note,
that using the DoMatchParen command might actually change the current
window and buffer.

So make sure, that after calling the matchparen command, the cursor
position has been restored