tar.vim: drop leading / in path names
A tar archive which containing files with leading `/` may cause
confusions as to where the content is extracted.  Let's make sure we
drop the leading `/` and use a relative path instead.

Also while at it, had to refactor it quite a bit and increase the
minimum supported Vim version to v9. Also add a test for some basic tar
functionality

Signed-off-by: Christian Brabandt <cb@256bit.org>